Share directory listing code among all platforms except androidNative
