+++ core/apple/src/files/FileSystemApple.kt
@@ -53,26 +53,6 @@ internal actual fun realpathImpl(path: String): String {
     }
 }
 
-
-@OptIn(ExperimentalForeignApi::class, ExperimentalStdlibApi::class)
-internal actual class OpaqueDirEntry constructor(private val dir: CPointer<DIR>) : AutoCloseable {
-    actual fun readdir(): String? {
-        val entry = readdir(dir) ?: return null
-        return entry[0].d_name.toKString()
-    }
-
-    override fun close() {
-        closedir(dir)
-    }
-}
-
-@OptIn(ExperimentalForeignApi::class)
-internal actual fun opendir(path: String): OpaqueDirEntry {
-    val dirent = platform.posix.opendir(path)
-    if (dirent != null) return OpaqueDirEntry(dirent)
-    throw IOException("Can't open directory $path: ${strerror(errno)?.toKString() ?: "reason unknown"}")
-}
-
 internal actual fun metadataOrNullImpl(path: Path): FileMetadata? {
     val attributes = NSFileManager.defaultManager().fileAttributesAtPath(path.path, traverseLink = true) ?: return null
     val fileType = attributes[NSFileType] as String
