+++ core/linux/src/files/FileSystemLinux.kt
@@ -5,9 +5,12 @@
 
 package kotlinx.io.files
 
-import kotlinx.cinterop.*
-import kotlinx.io.IOException
-import platform.posix.*
+import kotlinx.cinterop.ExperimentalForeignApi
+import kotlinx.cinterop.cstr
+import kotlinx.cinterop.memScoped
+import kotlinx.cinterop.toKString
+import platform.posix.__xpg_basename
+import platform.posix.dirname
 
 @OptIn(ExperimentalForeignApi::class)
 internal actual fun dirnameImpl(path: String): String {
@@ -27,22 +30,3 @@ internal actual fun basenameImpl(path: String): String {
 }
 
 internal actual fun isAbsoluteImpl(path: String): Boolean = path.startsWith('/')
-
-@OptIn(ExperimentalForeignApi::class, ExperimentalStdlibApi::class)
-internal actual class OpaqueDirEntry constructor(private val dir: CPointer<DIR>) : AutoCloseable {
-    actual fun readdir(): String? {
-        val entry = readdir(dir) ?: return null
-        return entry[0].d_name.toKString()
-    }
-
-    override fun close() {
-        closedir(dir)
-    }
-}
-
-@OptIn(ExperimentalForeignApi::class)
-internal actual fun opendir(path: String): OpaqueDirEntry {
-    val dirent = platform.posix.opendir(path)
-    if (dirent != null) return OpaqueDirEntry(dirent)
-    throw IOException("Can't open directory $path: ${strerror(errno)?.toKString() ?: "reason unknown"}")
-}
