+++ core/mingw/src/files/FileSystemMingw.kt
@@ -60,23 +60,3 @@ internal actual fun realpathImpl(path: String): String {
         return buffer.toKString()
     }
 }
-
-
-@OptIn(ExperimentalForeignApi::class, ExperimentalStdlibApi::class)
-internal actual class OpaqueDirEntry constructor(private val dir: CPointer<DIR>) : AutoCloseable {
-    actual fun readdir(): String? {
-        val entry = readdir(dir) ?: return null
-        return entry[0].d_name.toKString()
-    }
-
-    override fun close() {
-        closedir(dir)
-    }
-}
-
-@OptIn(ExperimentalForeignApi::class)
-internal actual fun opendir(path: String): OpaqueDirEntry {
-    val dirent = platform.posix.opendir(path)
-    if (dirent != null) return OpaqueDirEntry(dirent)
-    throw IOException("Can't open directory $path: ${strerror(errno)?.toKString() ?: "reason unknown"}")
-}
